Move editor toolbar initialization to a separate class EditorToolbarInitializer.java
